requirements.txt
fastapi
uvicorn
pinecone
sentence-transformers
requests
python-dotenv
langchain